Lift raw_socket_stream implementation out into own files
